Upscaling: First Generate Then Upscale
The script now generates the low-res images first and then performs an upscaling function. This is good because you are less likely to avoid memory crashes when there is a large batch and more iterations. This way the memory is not shared between SD and GFPGAN.

Probably the best solution to this particular issue.