Refactored Viewer
No longer a modal. But instead replaces the tabs. Also fixed the context menu to close the viewer and go to the needed tabs accordingly. The lag should be fixed. The transition is not fully smooth because the gallery is loaded again .. The only way to optimize this really is to get thumbnails done.