make --autoconvert option import .safetensors too
Since 2.3.0 launching `invokeai` with `--autoconvert /path/to/directory`
would recursively search for `.ckpt` files in the directory, convert them
into `diffusers` models, and import into `models.yaml`. This PR adds
support for `.safetensors` as well.

Note that we use brittle heuristics to distinguish inpainting models
from v1.x, from v2.x models -- if the model must contain 'inpaint' in
its name or path, it is treated as a 1.5 inpainting model. If the key
"model.diffusion_model.input_blocks.2.1.transformer_blocks.0.attn2.to_k.weight"
is present in the checkpoint file and has a shape ending in 1024, then
it is treated as a 2.x model.

The model's original VAE is always used. The VAE can be edited in the
Web GUI and CLI, but possibly the VAE for all 1.X-derived models
should be set to stability/sd-ft-vae. (comments welcome)