Podman: Use $CONTAINER_ENGINE env for "podman" instead of default "docker"
build.sh:

When creating the volume, grab its path and do a "podman unshare"
to give the running "appuser" access to the volume's files.  Without it, you'd
have to run the rootless user as "root" in the container, and no one wants
that.

run.sh:

"--platform" seems to have problems with podman. This may actually
be an issue with arm64 as I'm trying both at the same time.  Regardless,
I unset it on podman to keep behavior the same for docker.