fix(nodes): alternative way to mark properties required
Note: I don't think this PR is the right way to do things. The existing method of using a `Config` class is the "official" solution. See:
https://docs.pydantic.dev/usage/schema/#schema-customization
https://github.com/pydantic/pydantic/discussions/4577

This shouldn't interfere with pydantic; we are following the blessed path to customize a schema.

That said, here's another way to accomplish the same thing:

Instead of adding a `Config` class with `schema_extra` props, we can manipulate the schema generation directly.

Iterate over the properties of each output model to build an array of required properties.

This method requires an output actually be used by an invocation, else it is not pulled in as an output. So `PromptOutput` needs an invocation for this method to work for it.