nodes-api: initialize web config on module import
this enables the use of a debugger (e.g. the VSCode debugger) which runs
the API app with the `uvicorn` server. Without this change, the FastAPI
app can only be run by executing api_app:__main__, and crashes on start
if ran with `uvicorn` because `web_config` is an empty dict at that point.